pyproject.toml
[tool.poetry]
name = "research-chain-v1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain-openai = "^0.1.6"
python-dotenv = "^1.0.1"
python-docx = "^1.1.2"
langchain = "^0.1.20"
bs4 = "^0.0.2"
selenium = "^4.20.0"
chromedriver = "^2.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
